float32 instead of float

+++ skdata/tests/test_caltech.py
@@ -4,7 +4,7 @@ from skdata import tasks
 
 def test_Caltech101():
     dset = caltech.Caltech101()
-    task = dset.img_classification_task(dtype='float')
+    task = dset.img_classification_task(dtype='float32')
     tasks.assert_img_classification(*task, N=9144)
 
     X, y = task
@@ -15,5 +15,5 @@ def test_Caltech101():
 
 def test_Caltech256():
     dset = caltech.Caltech256()
-    task = dset.img_classification_task(dtype='float')
+    task = dset.img_classification_task(dtype='float32')
     tasks.assert_img_classification(*task)
